+++ src/tests/integration/components/modals/clean-obsolete-tokens-modal-test.js
@@ -210,8 +210,7 @@ describe('Integration | Component | modals/clean obsolete tokens modal', functio
       .then(() => click($submitButton[0]))
       .then(() => {
         expect($submitButton).to.have.class('in-flight');
-        expect(submitStub.lastCall.args[0].toArray())
-          .to.deep.equal(this.get('tokens').toArray());
+        expect(submitStub.lastCall.args[0]).to.have.same.members(this.get('tokens'));
       });
   });
 
@@ -227,8 +226,8 @@ describe('Integration | Component | modals/clean obsolete tokens modal', functio
       .then(() => click($submitButton[0]))
       .then(() => {
         expect($submitButton).to.have.class('in-flight');
-        expect(submitStub.lastCall.args[0].toArray())
-          .to.deep.equal(tokens.without(tokens[3]));
+        expect(submitStub.lastCall.args[0])
+          .to.have.same.members(tokens.without(tokens[3]));
       });
   });